arduino.asli: configure i2c-6 mux/dir according to hardware guide
Settings for Arduino pinmux follow Table 4, where former 204 and 205 correspond to
MUX7_SEL and MUX5_SEL resp. and are set to L. Note: settings for these in Sec. 11.6
(H) are wrong.

Settings for 14 and 165 direction follow Table 7, where former 236 and 237 correspond to
MUX6_DIR and MUX4_DIR resp. and are set L to set to input (HiZ).

Signed-off-by: Ferry Toth <ftoth@exalondelft.nl>